Some minor fixes
Removed reboot on 0 memstate. Sometimes GPU gets memstate 0 but it is not actually crashed. Want to avoid unnecessary reboot. Also added 75 seconds sleep when script executes. I noticed in logs on hourly report that script executes twice. It is possible it is because script readds itself to cron each time it is executed so it is possible cron executes it twice.